pequenos fixes
