+++ routes/plataformas.router.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const axios = require("axios");
-const controller = require("../controllers/plataforma.controller");
-
+// const axios = require("axios");
+const controller = require("../controllers/plataformas.controller");
 
 router.get("/", controller.index);
 router.get("/:id", controller.show);
@@ -12,13 +11,13 @@ router.delete("/:id", controller.destroy);
 
 // Preguntar por la api de ubisoft
 //router.get("/:id/ubisoft", async (req, res) => {
-    //const { id } = req.params;
-    //try {
-        //const response = await axios.get(`https://api.ubisoft.com/games/${id}`);
-        //res.json(response.data);
-    //} catch (error) {
-        //res.status(500).json({ error: "Error al conectar con Ubisoft" });
-    //}
+//const { id } = req.params;
+//try {
+//const response = await axios.get(`https://api.ubisoft.com/games/${id}`);
+//res.json(response.data);
+//} catch (error) {
+//res.status(500).json({ error: "Error al conectar con Ubisoft" });
+//}
 //});
 
 module.exports = router;
